+++ protocols/application/ftp.h
@@ -14,6 +14,6 @@
 #define FTP_HEADER_MAX_LEN      512
 
 
-protocol_info dissect_ftp(const uint8_t *pkt, uint32_t pkt_len, output_format fmt);
+protocol_info dissect_ftp(const uint8_t *pkt, size_t pkt_len, output_format fmt);
 
 #endif
