+++ protocols/application/telnet.h
@@ -16,6 +16,6 @@
 #define TELNET_EXTENDED_OPT_LIST    255
 
 
-protocol_info dissect_telnet(const uint8_t *pkt, uint32_t pkt_len, output_format fmt);
+protocol_info dissect_telnet(const uint8_t *pkt, size_t pkt_len, output_format fmt);
 
 #endif
