+++ protocols/transport/udp.h
@@ -18,6 +18,6 @@
 #define UDP_LEN                 8
 
 
-protocol_info dissect_udp(const uint8_t *pkt, uint32_t pkt_len, output_format fmt);
+protocol_info dissect_udp(const uint8_t *pkt, size_t pkt_len, output_format fmt);
 
 #endif
