+++ utils/protocol.h
@@ -29,7 +29,7 @@ typedef struct protocol_info {
 typedef struct protocol_handler {
     int protocol;
 	protocol_layer layer;
-    protocol_info (*dissect_proto)(const uint8_t *pkt, uint32_t pkt_len, output_format fmt);
+    protocol_info (*dissect_proto)(const uint8_t *pkt, size_t pkt_len, output_format fmt);
     const char *protocol_name;
 } protocol_handler;
 
@@ -43,19 +43,19 @@ typedef struct protocol_handler_mapping {
 #define IS_NULL_HANDLER(hdl)	(0 == hdl.protocol && PROTOCOL_LAYER_NONE == hdl.layer && NULL == hdl.dissect_proto && NULL == hdl.protocol_name)
 #define SHOW_OUTPUT(pkt, len, fmt, print_func, visualize_func) \
 		do { \
-			void (*output_func)(const uint8_t *, uint32_t) = select_output_func(fmt, print_func, visualize_func); \
+			void (*output_func)(const uint8_t *, size_t) = select_output_func(fmt, print_func, visualize_func); \
     		if (NULL != output_func && len > 0) output_func(pkt, len); \
 		} while(0)
 
 void *select_output_func(
     output_format fmt,
-    void (*print_func)(const uint8_t *, uint32_t), 
-    void (*visualize_func)(const uint8_t *, uint32_t)
+    void (*print_func)(const uint8_t *, size_t), 
+    void (*visualize_func)(const uint8_t *, size_t)
 );
 protocol_handler *create_protocol_handler(
 	int proto, 
 	protocol_layer layer, 
-	protocol_info (*dissect_proto)(const uint8_t *pkt, uint32_t pkt_len, output_format fmt),
+	protocol_info (*dissect_proto)(const uint8_t *pkt, size_t pkt_len, output_format fmt),
 	const char *protocol_name
 );
 protocol_handler_mapping *create_protocol_handler_mapping(
@@ -66,6 +66,5 @@ protocol_handler_mapping **create_mappings_arr();
 void add_mapping(protocol_handler_mapping ***arr_ptr, protocol_handler_mapping *new_mapping);
 void destroy_mappings(protocol_handler_mapping **mappings);
 protocol_handler get_protocol_handler(int target_proto, protocol_handler *proto_table);
-
 
 #endif
